+++ db/migrate/20250408114821_add_post_gis_extension.rb
@@ -1,0 +1,9 @@
+class AddPostGisExtension < ActiveRecord::Migration[6.1]
+  def up
+    enable_extension 'postgis'
+  end
+
+  def down
+    disable_extension 'postgis'
+  end
+end